Modified structure to follow Application Factory design pattern.
Using Blueprints to easier modularize application components, specially background tasks and their views.
