+++ Procfile
@@ -1,2 +1,2 @@
-web: gunicorn --worker-class eventlet -w 1 pbapi:core.app
-worker: celery worker -A pbapi.core.celery --concurrency=2 --pool=eventlet --loglevel=info
+web: gunicorn --worker-class eventlet -w 1 run.app
+worker: celery worker -A run.celery --concurrency=2 --pool=eventlet --loglevel=info
