+++ app/__init__.py
@@ -1,0 +1,3 @@
+from celery import Celery
+
+celery = Celery('app.factory')
