--- pbapi/__init__.py
@@ -1,1 +1,0 @@
-from pbapi.core import app
